+++ test/test_plugin-son-mano-smr.sh
@@ -28,30 +28,52 @@
 
 
 trap "set +e;
-docker rm -fv test.broker;
-docker rm -fv test.mongo;
-docker rm -fv test.pluginmanager
-docker rm -fv test.smr;
-docker rm -fv ssmexample" INT TERM EXIT
+if ! [[ "$(docker inspect -f {{.State.Running}} test.broker 2> /dev/null)" == "" ]]; then docker rm -fv test.broker ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.mongo 2> /dev/null)" == "" ]]; then docker rm -fv test.mongo ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.pluginmanager 2> /dev/null)" == "" ]]; then docker rm -fv test.pluginmanager ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.smr 2> /dev/null)" == "" ]]; then docker rm -fv test.smr ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} ssmexample 2> /dev/null)" == "" ]]; then docker rm -fv ssmexample ; fi
+# ensure f/ssms cleanup
+if ! [[ "$(docker inspect -f {{.State.Running}} ssmexample 2> /dev/null)" == "" ]]; then docker rm -fv ssmexample ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1firewallconfiguration1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1firewallconfiguration1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1dumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1dumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1monitoring1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1monitoring1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1updateddumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1updateddumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonssmservice1dumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonssmservice1dumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonssmservice1placement1 2> /dev/null)" == "" ]]; then docker rm -fv sonssmservice1placement1 ; fi
+echo end." INT TERM EXIT
 #docker network rm test.sonata-plugins" INT TERM EXIT
 
 # ensure cleanup
 set +e
-docker rm -fv test.broker
-docker rm -fv test.mongo
-docker rm -fv test.pluginmanager
-docker rm -fv test.smr
-
+if ! [[ "$(docker inspect -f {{.State.Running}} test.broker 2> /dev/null)" == "" ]]; then docker rm -fv test.broker ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.mongo 2> /dev/null)" == "" ]]; then docker rm -fv test.mongo ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.pluginmanager 2> /dev/null)" == "" ]]; then docker rm -fv test.pluginmanager ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.smr 2> /dev/null)" == "" ]]; then docker rm -fv test.smr ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} ssmexample 2> /dev/null)" == "" ]]; then docker rm -fv ssmexample ; fi
 # ensure f/ssms cleanup
-set +e
-docker rm -fv ssmexample
-docker rm -fv sonfsmservice1firewallconfiguration1
-docker rm -fv sonfsmservice1function1dumb1
-docker rm -fv sonfsmservice1function1monitoring1
-docker rm -fv sonfsmservice1function1updateddumb1
-docker rm -fv sonssmservice1dumb1
-docker rm -fv sonssmservice1placement1
-docker rm -fv sonssmservice1dumb1
+if ! [[ "$(docker inspect -f {{.State.Running}} ssmexample 2> /dev/null)" == "" ]]; then docker rm -fv ssmexample ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1firewallconfiguration1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1firewallconfiguration1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1dumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1dumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1monitoring1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1monitoring1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonfsmservice1function1updateddumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonfsmservice1function1updateddumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonssmservice1dumb1 2> /dev/null)" == "" ]]; then docker rm -fv sonssmservice1dumb1 ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} sonssmservice1placement1 2> /dev/null)" == "" ]]; then docker rm -fv sonssmservice1placement1 ; fi
+
+#docker rm -fv test.broker
+#docker rm -fv test.mongo
+#docker rm -fv test.pluginmanager
+#docker rm -fv test.smr
+# ensure f/ssms cleanup
+#set +e
+#docker rm -fv ssmexample
+#docker rm -fv sonfsmservice1firewallconfiguration1
+#docker rm -fv sonfsmservice1function1dumb1
+#docker rm -fv sonfsmservice1function1monitoring1
+#docker rm -fv sonfsmservice1function1updateddumb1
+#docker rm -fv sonssmservice1dumb1
+#docker rm -fv sonssmservice1placement1
+#docker rm -fv sonssmservice1dumb1
 
 #  always abort if an error occurs
 set -e
@@ -85,4 +107,7 @@ docker run --name test.smr --net=test.sonata-plugins --net-alias==specificmanage
 -v '/var/run/docker.sock:/var/run/docker.sock' \
 -e network_id=test.sonata-plugins  -e broker_man_host=http://broker:15672/ registry.sonata-nfv.eu:5000/specificmanagerregistry py.test -v
 
-echo "done."
+# Show containers logs
+docker logs test.broker test.mongo test.pluginmanager test.smr
+
+echo "done. #test_plugin-son-mano-smr"
