+++ test/test_son-mano-base.sh
@@ -36,16 +36,25 @@
 #
 
 trap "set +e
-docker rm -fv test.broker
-docker rm -fv test.sonmanobase" INT TERM EXIT
+if ! [[ "$(docker inspect -f {{.State.Running}} test.broker 2> /dev/null)" == "" ]]; then docker rm -fv test.broker ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.sonmanobase 2> /dev/null)" == "" ]]; then docker rm -fv test.sonmanobase ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.mongo 2> /dev/null)" == "" ]]; then docker rm -fv test.mongo ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.pluginmanager 2> /dev/null)" == "" ]]; then docker rm -fv test.pluginmanager ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.placementexecutive 2> /dev/null)" == "" ]]; then docker rm -fv test.placementexecutive ; fi
+echo done trap" INT TERM EXIT
 #docker network rm test.sonata-plugins" INT TERM EXIT
 
 # ensure cleanup
 set +e
-docker rm -fv test.broker
-docker rm -fv test.mongo
-docker rm -fv test.pluginmanager
-docker rm -fv test.placementexecutive
+if ! [[ "$(docker inspect -f {{.State.Running}} test.broker 2> /dev/null)" == "" ]]; then docker rm -fv test.broker ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.sonmanobase 2> /dev/null)" == "" ]]; then docker rm -fv test.sonmanobase ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.mongo 2> /dev/null)" == "" ]]; then docker rm -fv test.mongo ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.pluginmanager 2> /dev/null)" == "" ]]; then docker rm -fv test.pluginmanager ; fi
+if ! [[ "$(docker inspect -f {{.State.Running}} test.placementexecutive 2> /dev/null)" == "" ]]; then docker rm -fv test.placementexecutive ; fi
+#docker rm -fv test.broker
+#docker rm -fv test.mongo
+#docker rm -fv test.pluginmanager
+#docker rm -fv test.placementexecutive
 #docker network rm test.sonata-plugins
 
 #  always abort if an error occurs
@@ -81,9 +90,9 @@ docker run --name test.sonmanobase --net=test.sonata-plugins --net-alias=sonmano
 -v '/var/run/docker.sock:/var/run/docker.sock' \
 registry.sonata-nfv.eu:5000/sonmanobase py.test -v
 
+docker logs test.broker test.mongo test.pluginmanager test.sonmanobase
 
-echo "done."
-
+echo "done. #test_son-mano-base"
 
 ## setup cleanup mechanism
 #trap "set +e; docker rm -fv test.broker; docker rm -fv test.sonmanobase" INT TERM EXIT
